workflows: identifier invalid
The workflow is not valid. .github/workflows/build.yml (Line: 22, Col: 3): The identifier 'good revision' is invalid. IDs may only contain alphanumeric characters, '_', and '-'. IDs must start with a letter or '_' and and must be less than 100 characters. .github/workflows/build.yml (Line: 34, Col: 3): The identifier '14.x' is invalid. IDs may only contain alphanumeric characters, '_', and '-'. IDs must start with a letter or '_' and and must be less than 100 characters.

Signed-off-by: XSans0 <xsansdroid@gmail.com>